requirements.txt
psycopg2-binary
uvicorn
fastapi
python-jose 
passlib
passlib[argon2]==1.7.4
sqlalchemy
pydantic[email]
requests
bcrypt
asyncpg
